requirements.txt
agate==1.6.1
agate-dbf==0.2.1
agate-excel==0.2.3
agate-sql==0.5.4
appdirs==1.4.3
astroid==2.3.3
astropy==4.0
atomicwrites==1.3.0
attr==0.3.1
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
black==19.10b0
bleach==3.1.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
confuse==1.0.0
csvkit==1.0.4
cycler==0.10.0
dbfread==2.0.7
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
et-xmlfile==1.0.1
google-api-core==1.16.0
google-auth==1.11.2
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
htmlmin==0.1.12
idna==2.8
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
isort==4.3.21
jdcal==1.4.1
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.3
jupyterlab-pygments==0.1.0
kaggle==1.5.6
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
leather==0.3.3
llvmlite==0.31.0
MarkupSafe==1.1.1
matplotlib==3.2.0rc3
mccabe==0.6.1
missingno==0.4.2
mistune==0.8.4
more-itertools==8.2.0
nbconvert==6.0.0a0
nbformat==5.0.4
networkx==2.4
notebook==6.0.3
numba==0.48.0
numpy==1.18.1
openpyxl==3.0.3
packaging==20.1
pandas==0.25.3
pandas-profiling==2.5.0
pandocfilters==1.4.2
parsedatetime==2.5
parso==0.6.1
pathspec==0.7.0
phik==0.9.9
pickleshare==0.7.5
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.3
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.5.2
pylint==2.4.4
pyparsing==2.4.6
pyrsistent==0.15.7
pytest==5.3.5
pytest-pylint==0.15.0
python-dateutil==2.8.1
python-dotenv==0.11.0
python-slugify==4.0.0
pytimeparse==1.1.8
pytz==2019.3
pywin32==227
pywinpty==0.5.7
PyYAML==5.3
pyzmq==18.1.1
regex==2020.1.8
requests==2.22.0
rsa==4.0
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
six==1.14.0
SQLAlchemy==1.3.13
tangled-up-in-unicode==0.0.3
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
toml==0.10.0
tornado==6.0.3
tqdm==4.42.0
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.24.3
visions==0.2.2
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.11.2
xlrd==1.2.0
zipp==3.0.0
